requirements.txt
streamlit
ultralytics
opencv-python
numpy
inference_sdk
supervision
inference
python-dotenv 